Change toolbar per fragment, fix chat scroll up
