+++ app/src/main/java/com/can_apps/showcase/MainActivity.kt
@@ -1,34 +1,23 @@
 package com.can_apps.showcase
 
-import android.content.Intent
 import android.content.res.Resources
 import android.os.Bundle
 import android.util.Log
-import android.view.Menu
-import android.view.MenuItem
 import androidx.appcompat.app.AppCompatActivity
-import androidx.drawerlayout.widget.DrawerLayout
 import androidx.navigation.NavController
-import androidx.navigation.findNavController
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.AppBarConfiguration
-import androidx.navigation.ui.navigateUp
-import androidx.navigation.ui.onNavDestinationSelected
 import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
-import com.google.android.gms.oss.licenses.OssLicensesMenuActivity
 import com.google.android.material.navigation.NavigationView
-import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
 
-    private lateinit var appBarConfiguration: AppBarConfiguration
+//    private lateinit var appBarConfiguration: AppBarConfiguration
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-
-        setSupportActionBar(toolbar)
 
         val host: NavHostFragment = supportFragmentManager
             .findFragmentById(R.id.my_nav_host_fragment) as NavHostFragment? ?: return
@@ -36,13 +25,13 @@ class MainActivity : AppCompatActivity() {
         val navController = host.navController
 
         // Add this when using drawer \/
-        val drawerLayout: DrawerLayout? = findViewById(R.id.drawer_layout)
-        appBarConfiguration = AppBarConfiguration(
-            setOf(R.id.home_dest), // R.id.deeplink_dest),
-            drawerLayout
-        )
+//        val drawerLayout: DrawerLayout? = findViewById(R.id.drawer_layout)
+//        appBarConfiguration = AppBarConfiguration(
+//            setOf(R.id.home_dest), // R.id.deeplink_dest),
+//            drawerLayout
+//        )
 
-        setupActionBar(navController, appBarConfiguration)
+//        setupActionBar(navController, appBarConfiguration)
         setupNavigationMenu(navController)
 //        setupBottomNavMenu(navController)
 
@@ -77,35 +66,9 @@ class MainActivity : AppCompatActivity() {
         setupActionBarWithNavController(navController, appBarConfig)
     }
 
-    override fun onCreateOptionsMenu(menu: Menu): Boolean {
-        val retValue = super.onCreateOptionsMenu(menu)
-        val navigationView = findViewById<NavigationView>(R.id.nav_view)
-        // The NavigationView already has these same navigation items, so we only add
-        // navigation items to the menu here if there isn't a NavigationView
-        if (navigationView == null) {
-            menuInflater.inflate(R.menu.overflow_menu, menu)
-            return true
-        }
-        return retValue
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean =
-        when (item.itemId) {
-            R.id.license_dest -> {
-                startActivity(Intent(this, OssLicensesMenuActivity::class.java))
-                true
-            }
-            // Have the NavigationUI look for an action or destination matching the menu
-            // item id and navigate there if found.
-            // Otherwise, bubble up to the parent.
-            else ->
-                item.onNavDestinationSelected(findNavController(R.id.my_nav_host_fragment)) ||
-                    super.onOptionsItemSelected(item)
-        }
-
-    override fun onSupportNavigateUp(): Boolean {
-        // Allows NavigationUI to support proper up navigation or the drawer layout
-        // drawer menu, depending on the situation
-        return findNavController(R.id.my_nav_host_fragment).navigateUp(appBarConfiguration)
-    }
+//    override fun onSupportNavigateUp(): Boolean {
+//        // Allows NavigationUI to support proper up navigation or the drawer layout
+//        // drawer menu, depending on the situation
+//        return findNavController(R.id.my_nav_host_fragment).navigateUp(appBarConfiguration)
+//    }
 }
