+++ app/src/main/res/layout-w960dp/activity_main.xml
@@ -22,21 +22,11 @@
         android:layout_toEndOf="@id/nav_view"
         android:background="?android:attr/listDivider" />
 
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_toEndOf="@id/nav_view"
-        android:background="@color/colorPrimary"
-        android:theme="@style/ThemeOverlay.MaterialComponents.Dark.ActionBar" />
-
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/my_nav_host_fragment"
         android:name="androidx.navigation.fragment.NavHostFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_below="@id/toolbar"
         android:layout_toEndOf="@id/nav_view"
         app:defaultNavHost="true"
         app:navGraph="@navigation/nav_graph_home" />
