+++ common/src/main/res/values/colors.xml
@@ -1,5 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
+    <color name="colorPrimary">#6200EE</color>
+    <color name="colorPrimaryDark">#3700B3</color>
+    <color name="colorAccent">#03DAC5</color>
     <color name="rank_title">#E47A00</color>
     <color name="rank_background">#FFF1C7</color>
     <color name="rank_subtitle">#FFFFFF</color>
